requirements.txt
Django==4.0.7
django-crispy-forms==1.14.0
djangorestframework==3.13.1
django-extensions==3.2.0
django-taggit==3.0.0
django-filter==2.2.0
Markdown==3.4.1
six==1.16.0